+++ widgets/RichText/src/index.js
@@ -1,0 +1,37 @@
+import React, { useState } from 'react';
+import BraftEditor from 'braft-editor';
+import 'braft-editor/dist/index.css';
+
+const { createEditorState } = BraftEditor;
+
+const RichTextEditor = ({ name, onChange, value, ...rest }) => {
+  const [editor, set] = useState(() => createEditorState(value));
+
+  const handleChange = editor => {
+    set(editor);
+    const htmlContent = editor.toHTML();
+    onChange(name, htmlContent);
+  };
+
+  const onSave = () => {
+    // Pressing ctrl + s when the editor has focus will execute this method
+    // Before the editor content is submitted to the server, you can directly call editorState.toHTML () to get the HTML content
+    // const htmlContent = editor.toHTML()
+    // console.log(htmlContent)
+    // onChange(name, htmlContent)
+  };
+
+  return (
+    <div style={{ border: '1px solid rgba(0,0,0,0.2)' }}>
+      <BraftEditor
+        contentStyle={{ minHeight: 100, height: 'auto', maxHeight: 500 }}
+        value={editor}
+        onChange={handleChange}
+        onSave={onSave}
+        onBlur={onSave}
+      />
+    </div>
+  );
+};
+
+export default RichTextEditor;