+++ widgets/template/src/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Template() {
+  return <div>hello world</div>;
+}